[core] [no-op] Fix refactor dependency issue (#49984)
Fix broken build on windows platform.

---------
