+++ src/ray/util/BUILD
@@ -81,6 +81,7 @@ ray_cc_library(
         "subreaper.cc",
     ],
     deps = [
+        ":cmd_line_utils",
         ":logging",
         ":macros",
         "@boost//:asio",
@@ -141,6 +142,7 @@ ray_cc_library(
 ray_cc_library(
     name = "string_utils",
     hdrs = ["string_utils.h"],
+    srcs = ["string_utils.cc"],
 )
 
 ray_cc_library(
@@ -176,6 +178,16 @@ ray_cc_library(
     hdrs = ["sample.h"],
 )
 
+ray_cc_library(
+    name = "cmd_line_utils",
+    hdrs = ["cmd_line_utils.h"],
+    srcs = ["cmd_line_utils.cc"],
+    deps = [
+        ":logging",
+        ":string_utils",
+    ],
+)
+
 # A giant 'util' target is split since PR https://github.com/ray-project/ray/pull/49938, here we keep the 'util' target for API compatibility.
 #
 # TODO(hjiang): We include a bunch of misc util function/class inside of the class, should split into multiple files and build targets.
@@ -185,6 +197,7 @@ ray_cc_library(
     srcs = ["util.cc"],
     deps = [
         ":array",
+        ":cmd_line_utils",
         ":container_util",
         ":counter_map",
         ":event",
