+++ src/ray/util/process.cc
@@ -30,14 +30,16 @@
 #include <unistd.h>
 #endif
 
-#include <string.h>
-
 #include <algorithm>
+#include <array>
 #include <atomic>
+#include <cstdio>
+#include <cstring>
 #include <fstream>
 #include <string>
 #include <vector>
 
+#include "ray/util/cmd_line_utils.h"
 #include "ray/util/filesystem.h"
 #include "ray/util/logging.h"
 #include "ray/util/macros.h"
