+++ src/ray/util/string_utils.cc
@@ -12,19 +12,30 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#pragma once
-
-#include <string>
+#include "ray/util/string_utils.h"
 
 namespace ray {
 
-// Transfer the string to the Hex format. It can be more readable than the ANSI mode
-std::string StringToHex(const std::string &str);
+std::string StringToHex(const std::string &str) {
+  constexpr char hex[] = "0123456789abcdef";
+  std::string result;
+  for (size_t i = 0; i < str.size(); i++) {
+    unsigned char val = str[i];
+    result.push_back(hex[val >> 4]);
+    result.push_back(hex[val & 0xf]);
+  }
+  return result;
+}
 
-/// Uses sscanf() to read a token matching from the string, advancing the iterator.
-/// \param c_str A string iterator that is dereferenceable. (i.e.: c_str < string::end())
-/// \param format The pattern. It must not produce any output. (e.g., use %*d, not %d.)
-/// \return The scanned prefix of the string, if any.
-std::string ScanToken(std::string::const_iterator &c_str, std::string format);
+std::string ScanToken(std::string::const_iterator &c_str, std::string format) {
+  int i = 0;
+  std::string result;
+  format += "%n";
+  if (static_cast<size_t>(sscanf(&*c_str, format.c_str(), &i)) <= 1) {
+    result.insert(result.end(), c_str, c_str + i);
+    c_str += i;
+  }
+  return result;
+}
 
 }  // namespace ray
